+++ app/controllers/application_controller.rb
@@ -2,7 +2,6 @@
 
 class ApplicationController < ActionController::Base
   include Pundit
-  include InvitableConcern
   protect_from_forgery
   before_action :configure_permitted_parameters, if: :devise_controller?
   before_action :store_user_location!, if: :storable_location?
