+++ app/models/rdv.rb
@@ -137,10 +137,9 @@ class Rdv < ApplicationRecord
     status.in? CANCELLED_STATUSES
   end
 
-  def cancellable?
+  def editable?
     !cancelled? && starts_at > 4.hours.from_now
   end
-  alias editable? cancellable?
 
   def available_to_file_attente?
     !cancelled? && starts_at > 7.days.from_now && !home?
