+++ app/policies/user/rdv_policy.rb
@@ -7,18 +7,20 @@ class User::RdvPolicy < ApplicationPolicy
     (record.user_ids & current_user.available_users_for_rdv.pluck(:id)).any?
   end
 
+  def index?
+    !current_user.only_invited?
+  end
+
   alias new? rdv_belongs_to_user_or_relatives?
   alias show? rdv_belongs_to_user_or_relatives?
   alias create? rdv_belongs_to_user_or_relatives?
 
-  # three next are used in creneaux controller when coming from file attentes
-  alias index? rdv_belongs_to_user_or_relatives?
-  alias edit? rdv_belongs_to_user_or_relatives?
-  alias update? rdv_belongs_to_user_or_relatives?
-
-  def cancel?
-    record.cancellable? && rdv_belongs_to_user_or_relatives?
+  def edit?
+    record.editable? && rdv_belongs_to_user_or_relatives?
   end
+  alias cancel? edit?
+  alias creneaux? edit?
+  alias update? edit?
 
   class Scope < Scope
     alias current_user pundit_user
