+++ app/views/layouts/application.html.slim
@@ -7,7 +7,7 @@ html lang="fr"
       = render "layouts/rdv_solidarites_instance_name"
       = render "layouts/degraded_service", message: ENV["DEGRADED_SERVICE_MESSAGE_USERS"]
       .container
-        - if current_user.present? && !@invited_user
+        - if current_user.present? && !current_user.only_invited?
           = render "common/header_user_logged"
         - else
           = render "common/header"
