+++ app/views/users/rdvs/_creneaux.html.slim
@@ -3,7 +3,7 @@ div id="creneaux-lieu"
     - previous_from_date = date_range.begin - 7.days
     - if date_range.begin > Time.zone.today
       .col-12.col-md-auto.mb-2.mb-md-0.d-flex.align-items-center.justify-content-center
-        = link_to users_creneaux_index_path(rdv_id: rdv.id, date: previous_from_date), remote: true, class: "btn btn-primary", data: { disable_with: "..." } do
+        = link_to creneaux_users_rdv_path(rdv, date: previous_from_date), remote: true, class: "btn btn-primary", data: { disable_with: "..." } do
           i.fa.fa-chevron-left
           span.d-md-none.ml-1<
             | sem. précédente
@@ -21,10 +21,10 @@ div id="creneaux-lieu"
             - if creneaux_for_date.any?
               - creneaux_for_date.each do |date, creneaux|
                 - creneaux.sort_by(&:starts_at).each do |creneau|
-                  = link_to l(creneau.starts_at, format: "%H:%M"), edit_users_creneaux_path(@rdv.id, starts_at: creneau.starts_at, agent_id: creneau.agent.id), class: "btn btn-light mr-1 mb-1 w-100"
+                  = link_to l(creneau.starts_at, format: "%H:%M"), edit_users_rdv_path(@rdv, starts_at: creneau.starts_at, agent_id: creneau.agent.id), class: "btn btn-light mr-1 mb-1 w-100"
     - if date_range.end < @all_creneaux.last.starts_at.to_date
       .col-12.col-md-auto.mt-2.mt-md-0.d-flex.align-items-center.justify-content-center
-        = link_to users_creneaux_index_path(rdv_id: rdv.id, date: date_range.end + 1.day), remote: true, class: "btn btn-primary", data: { disable_with: "..." } do
+        = link_to creneaux_users_rdv_path(rdv, date: date_range.end + 1.day), remote: true, class: "btn btn-primary", data: { disable_with: "..." } do
           span.d-md-none.mr-1
             | sem. prochaine
           i.fa.fa-chevron-right
