+++ app/views/users/rdvs/creneaux.js.erb
@@ -5,4 +5,3 @@
     element.innerHTML = "<%= escape_javascript creneaux_html %>";
   }
 })();
-
