+++ app/views/users/rdvs/edit.html.slim
@@ -8,6 +8,6 @@
       h4 Nouveau créneau selectionné
 
       p= l(@starts_at, format: :human).capitalize
-      = link_to "Confirmer le nouveau créneau", users_creneaux_path(@rdv.id, starts_at: @starts_at.to_s), class:"btn btn-primary", method: :put
+      = link_to "Confirmer le nouveau créneau", users_rdv_path(@rdv, starts_at: @starts_at.to_s), class:"btn btn-primary", method: :put
       br
-      = link_to "Retour à la liste", users_creneaux_index_path(rdv_id: @rdv.id), class:"btn"
+      = link_to "Retour à la liste", creneaux_users_rdv_path(@rdv), class:"btn"
