+++ spec/controllers/users/rdvs_controller_spec.rb
@@ -62,7 +62,7 @@ RSpec.describe Users::RdvsController, type: :controller do
         travel_to(now)
         sign_in rdv.users.first
 
-        put :cancel, params: { rdv_id: rdv.id }
+        put :cancel, params: { id: rdv.id }
         expect(rdv.reload.cancelled_at).to be_within(5.seconds).of(now)
       end
 
@@ -70,13 +70,13 @@ RSpec.describe Users::RdvsController, type: :controller do
         rdv = create(:rdv, starts_at: 5.hours.from_now)
         sign_in rdv.users.first
         expect(RdvUpdater).to receive(:update).with(rdv.users.first, rdv, { status: "excused" })
-        put :cancel, params: { rdv_id: rdv.id }
+        put :cancel, params: { id: rdv.id }
       end
 
-      it "redirects to the rrdv" do
+      it "redirects to the rdv" do
         rdv = create(:rdv, starts_at: 5.hours.from_now)
         sign_in rdv.users.first
-        put :cancel, params: { rdv_id: rdv.id }
+        put :cancel, params: { id: rdv.id }
         expect(response).to redirect_to users_rdv_path(rdv)
       end
 
@@ -84,7 +84,7 @@ RSpec.describe Users::RdvsController, type: :controller do
         rdv = create(:rdv, starts_at: 3.hours.from_now)
         sign_in rdv.users.first
         expect do
-          put :cancel, params: { rdv_id: rdv.id }
+          put :cancel, params: { id: rdv.id }
         end.not_to change(rdv, :cancelled_at)
       end
     end
@@ -96,7 +96,7 @@ RSpec.describe Users::RdvsController, type: :controller do
       sign_in other_user
 
       expect do
-        put :cancel, params: { rdv_id: rdv.id }
+        put :cancel, params: { id: rdv.id }
       end.to raise_error(ActiveRecord::RecordNotFound)
     end
   end
@@ -157,6 +157,194 @@ RSpec.describe Users::RdvsController, type: :controller do
           expect(response.body).to match(/Le mardi 20 octobre 2020/)
           expect(response.body).to match(/Modifier l&#39;horaire du RDV/)
         end
+      end
+    end
+  end
+
+  describe "GET #index" do
+    subject { get :index }
+
+    let!(:user) { create(:user) }
+    let!(:rdv1) { create(:rdv, users: [user], starts_at:  5.days.from_now) }
+    let!(:rdv2) { create(:rdv, users: [user], starts_at:  4.days.from_now) }
+
+    context "when signed in" do
+      before { sign_in user }
+
+      it "lists the rdvs" do
+        subject
+
+        expect(response).to be_successful
+        expect(response.body).to include(I18n.l(rdv1.starts_at, format: :human).to_s)
+        expect(response.body).to include(I18n.l(rdv2.starts_at, format: :human).to_s)
+      end
+    end
+
+    context "when not signed in" do
+      it "redirects to sign in path" do
+        subject
+
+        expect(response).to redirect_to(new_user_registration_path)
+      end
+
+      context "with a valid invitation token" do
+        let!(:invitation_token) do
+          user.invite! { |u| u.skip_invitation = true }
+          user.raw_invitation_token
+        end
+
+        it "is not authorized" do
+          get :index, params: { invitation_token: invitation_token }
+
+          expect(response).to redirect_to(root_path)
+          expect(flash[:error]).to eq("Vous ne pouvez pas effectuer cette action.")
+        end
+      end
+    end
+  end
+
+  describe "GET #creneaux" do
+    subject do
+      get :creneaux, params: { id: rdv.id }
+      rdv.reload
+    end
+
+    let(:organisation) { create(:organisation) }
+    let(:now) { "01/01/2019 10:00".to_datetime }
+    let!(:agent) { create(:agent, basic_role_in_organisations: [organisation]) }
+    let!(:lieu) { create(:lieu, address: "10 rue de la Ferronerie 44100 Nantes", organisation: organisation) }
+    let!(:motif) { create(:motif, organisation: organisation, max_booking_delay: 2.weeks.to_i) }
+    let!(:user) { create(:user) }
+    let(:rdv) { create(:rdv, users: [user], starts_at: 5.days.from_now, lieu: lieu, motif: motif, organisation: organisation) }
+
+    before do
+      travel_to(now)
+      sign_in user
+    end
+
+    context "no creneaux available" do
+      before { subject }
+
+      it { expect(assigns(:all_creneaux)).to be_empty }
+      it { expect(response.body).to include("Malheureusement, tous les créneaux sont pris.") }
+    end
+
+    context "creneaux available" do
+      let!(:plage_ouverture) do
+        create(:plage_ouverture, :daily, first_day: now + 3.days, start_time: Tod::TimeOfDay.new(10), lieu: lieu, agent: agent, motifs: [motif], organisation: organisation)
+      end
+
+      before { subject }
+
+      it { expect(response.body).to include("Voici les créneaux disponibles pour modifier votre rendez-vous du") }
+      it { expect(response.body).to include(I18n.l(rdv.starts_at, format: :human).to_s) }
+      it { expect(assigns(:date_range)).to eq(3.days.from_now.to_date..9.days.from_now.to_date) }
+      it { expect(assigns(:creneaux)).not_to be_empty }
+    end
+
+    context "when the rdv cannot be edited" do
+      let(:rdv) { create(:rdv, users: [user], starts_at: 2.days.ago, lieu: lieu, motif: motif, organisation: organisation) }
+
+      before { subject }
+
+      it { expect(response).to redirect_to(authenticated_user_root_path) }
+      it { expect(flash[:error]).to eq("Vous ne pouvez pas effectuer cette action.") }
+    end
+  end
+
+  describe "GET #edit" do
+    subject do
+      get :edit, params: { id: rdv.id, starts_at: starts_at }
+      rdv.reload
+    end
+
+    let(:organisation) { create(:organisation) }
+    let(:starts_at) { 3.days.from_now }
+    let(:now) { "01/01/2019 10:00".to_datetime }
+    let!(:agent) { create(:agent, basic_role_in_organisations: [organisation]) }
+    let!(:lieu) { create(:lieu, address: "10 rue de la Ferronerie 44100 Nantes", organisation: organisation) }
+    let!(:motif) { create(:motif, organisation: organisation) }
+    let!(:user) { create(:user) }
+    let(:rdv) { create(:rdv, users: [user], starts_at: 5.days.from_now, lieu: lieu, motif: motif, organisation: organisation) }
+
+    before do
+      travel_to(now)
+      sign_in user
+
+      allow(Users::CreneauSearch).to receive(:creneau_for)
+        .with(user: user, starts_at: starts_at, motif: motif, lieu: lieu)
+        .and_return(returned_creneau)
+    end
+
+    context "creneau is available" do
+      let(:returned_creneau) { Creneau.new }
+
+      before { subject }
+
+      it { expect(response.body).to include("Modification du Rendez-vous") }
+      it { expect(response.body).to include("Confirmer le nouveau créneau") }
+    end
+
+    context "creneau isn't available" do
+      let(:returned_creneau) { nil }
+
+      before { subject }
+
+      it { expect(response).to redirect_to(creneaux_users_rdv_path(rdv)) }
+    end
+  end
+
+  describe "PUT #update" do
+    let(:organisation) { create(:organisation) }
+    let(:now) { Time.zone.parse("01/01/2019 10:00") }
+    let(:starts_at) { 3.days.from_now }
+    let(:user) { create(:user) }
+    let(:motif) { create(:motif, organisation: organisation) }
+    let(:lieu) { create(:lieu, address: "10 rue de la Ferronerie 44100 Nantes", organisation: organisation) }
+    let!(:agent) { create(:agent, basic_role_in_organisations: [organisation]) }
+    let(:rdv) { create(:rdv, users: [user], starts_at: 5.days.from_now, lieu: lieu, motif: motif, organisation: organisation) }
+
+    before do
+      travel_to(now)
+      sign_in user
+      allow(Users::CreneauSearch).to receive(:creneau_for)
+        .with(user: user, starts_at: starts_at, motif: motif, lieu: lieu)
+        .and_return(returned_creneau)
+    end
+
+    context "with an available creneau" do
+      let(:returned_creneau) { Creneau.new(starts_at: starts_at, agent: agent) }
+
+      it "respond success and update RDV" do
+        put :update, params: { id: rdv.id, starts_at: starts_at, agent_id: agent.id }
+        expect(response).to redirect_to(users_rdv_path(rdv))
+        expect(flash[:success]).to eq("Votre RDV a bien été modifié")
+        expect(rdv.reload.starts_at).to eq(starts_at)
+        expect(rdv.reload.agent_ids).to eq([agent.id])
+        expect(rdv.reload.created_by).to eq("file_attente")
+      end
+
+      context "when it cannot be updated" do
+        before do
+          allow_any_instance_of(Rdv).to receive(:update).and_return(false)
+        end
+
+        it "redirects to creneaux index" do
+          put :update, params: { id: rdv.id, starts_at: starts_at, agent_id: agent.id }
+          expect(response).to redirect_to(creneaux_users_rdv_path(rdv))
+          expect(flash[:error]).to eq("Le RDV n'a pas pu être modifié")
+        end
+      end
+    end
+
+    context "without an available creneau" do
+      let(:returned_creneau) { nil }
+
+      it "redirect to index when not available" do
+        put :update, params: { id: rdv.id, starts_at: starts_at, agent_id: agent.id }
+
+        expect(response).to redirect_to(creneaux_users_rdv_path(rdv))
+        expect(flash[:alert]).to eq("Ce créneau n'est plus disponible")
       end
     end
   end
